refactor(circuit): split dense-matrix helpers out of circuit_utils [Clean Code: SRP]
Move to a new `scqubits/core/dense_matrix_helpers.py`:

- `_cos_dia`, `_sin_dia` (sparse-output diagonal cos/sin)
- `_cos_dia_dense`, `_sin_dia_dense` (dense-output variants)
- `matrix_power_sparse`

The four `_*_dia*` symbols are private — internal callers
(`circuit_routines.py`, `tests/test_circuit_utils.py`) import directly
from the new module; no shim. `matrix_power_sparse` is public, so
keep a `from scqubits.core.dense_matrix_helpers import
matrix_power_sparse  # noqa: F401` deprecation shim in
`circuit_utils.py`.
