+++ scqubits/core/circuit_routines.py
@@ -52,18 +52,20 @@ from scqubits.core.charge_basis_operators import (
     _sin_theta,
 )
 from scqubits.core.circuit_utils import (
-    _cos_dia,
-    _cos_dia_dense,
     _generate_symbols_list,
-    _sin_dia,
-    _sin_dia_dense,
     get_trailing_number,
     grid_operator_func_factory,
     hierarchical_diagonalization_func_factory,
-    matrix_power_sparse,
     operator_func_factory,
     round_symbolic_expr,
     sawtooth_potential,
+)
+from scqubits.core.dense_matrix_helpers import (
+    _cos_dia,
+    _cos_dia_dense,
+    _sin_dia,
+    _sin_dia_dense,
+    matrix_power_sparse,
 )
 from scqubits.core.discretized_phi_operators import (
     _cos_phi,
